backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
python-dotenv
httpx
anthropic
